requirements.txt
numpy
torch
torchvision
tqdm
Pillow
matplotlib
scikit-learn
scikit-image
pandas
seaborn
opencv-python
scipy
kagglehub